Implementação RN: Cadastrar Medicação e ajustes

+++ NAC1-WS-RemedioManager/src/br/com/fiap/dao/PessoaDAO.java
@@ -2,6 +2,6 @@ package br.com.fiap.dao;
 
 import br.com.fiap.entity.Pessoa;
 
-public interface PessoaDAO extends GenericDAO<Pessoa, Integer>{
+public interface PessoaDAO extends GenericDAO<Pessoa, Long>{
 
 }
