+++ NAC1-WS-RemedioManager/src/br/com/fiap/dao/impl/PessoaDAOImpl.java
@@ -5,7 +5,7 @@ import javax.persistence.EntityManager;
 import br.com.fiap.dao.PessoaDAO;
 import br.com.fiap.entity.Pessoa;
 
-public class PessoaDAOImpl extends GenericDAOImpl<Pessoa, Integer> implements PessoaDAO{
+public class PessoaDAOImpl extends GenericDAOImpl<Pessoa, Long> implements PessoaDAO{
 
 	public PessoaDAOImpl(EntityManager em) {
 		super(em);
